refactor: cyclic dependencies 제거
