+++ src/main/java/nextstep/ladder/domain/factory/LadderFactoryBean.java
@@ -1,0 +1,14 @@
+package nextstep.ladder.domain.factory;
+
+import nextstep.ladder.domain.impl.board.LadderBoardCreator;
+import nextstep.ladder.domain.engine.BoardCreator;
+import nextstep.ladder.domain.engine.LineCreator;
+import nextstep.ladder.domain.impl.line.RandomLineCreator;
+
+public class LadderFactoryBean {
+
+    public static BoardCreator createLadderFactory() {
+        LineCreator lineCreator = new RandomLineCreator();
+        return new LadderBoardCreator(lineCreator);
+    }
+}
